+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * It is now possible to register contextual dependencies for class methods in the container.
 * Added a HTTP status code helper class.
 * Added missing status codes to `Response` class.
+* Added a `Retry` helper class that allows you to retry a callable a set number of times.
 
 --------------------------------------------------------